ok, we are in the 3rd round. and the former judge how has 10 cards.
so -i dont know how you have it set to work, but the way i got it to work in a vanilla version was to have all players get 5 cards dealt to them when the game started, but the player who was designated as judge just had their hand go invisible. that way when the next round started, it just became visible to them and play could carry on. i think it kept the card logic more simple because it was the same for every player until they were a judge when their hand became invisible.