ok i think that it all worked!
I have two small style changes i would like to make. on that overlay screen can we change it to say "Loading the terribleness..." instead?

also on the ready button, can this say "tap when ready" since most people are going to be using a phone to interact with this game.