keep the carousel, continue with the lobby problem.
note - if i refresh the either screen, that usually puts it in the right place. meaning if i were a player and i refreshed the screen/browser, it would put me in the right place to continue the game. just wanted you to have that information in case it helped. ask me for any clarification you need.