i love all of this - keep it in mind, especially the user tutorial.
for now i think i want to try some small styling and UI changes.

the response cards in a player's hand are hard to see- the border is set to go to black when the mouse cursor rolls over it, but since most people will be playing this on a phone, and without a mouse, can we darken those response card borders to be more visible? maybe an 18% grey or something? i like what they look like when they are selected. dont change that.