yes, i think your plan is on the right path. i also think this seems like the hard part. dont code anything yet
as far as the UI goes, i think the judge should have some kind of pop up modal thing that asks them if they want to add the winning response to the deck (or some wording like that, and just give them a yes or no button like the ones at the end of the game).

this should happen before the round/game announcement  so if doesn't interfere with that logic.

does that make sense?
 dont code anything yet.

ALSO - quick style change for the write your own card- kill the emoji at the end.