ok you also missed the more important part of my message -
when i was adding players to the lobby, i added nick on the left first, selected an avatar (not the first one) and hit join game. when i did the same to player becky, on the right, i tried to select an avatar and when it swtiched, it just added becky to the lobby. is this going to break everything again?

if so, im not sure i want to deal with this right this minute as its late and ive been doing this for a while now.