ok that seemed to work to stop the loop - but there are some things to address. the yellow game in progress box that mentions the game phase. this needs to be something else.
also, it says wait for the next round or reset.  when there is a new round, that player isnt offered a chance to join. does there need to be some kind of logic that looks for entering players between rounds to let them in? should we just make players wait until the next game?